move bin back to JS
